use sequence numbers instead of timeout for bidirectional neighbour detection
git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@305 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ originator.c
@@ -90,8 +90,8 @@ struct orig_node *get_orig_node( uint32_t addr ) {
 	orig_node->router = NULL;
 	orig_node->batman_if = NULL;
 
-	orig_node->bidirect_link = debugMalloc( found_ifs * sizeof(uint32_t), 402 );
-	memset( orig_node->bidirect_link, 0, found_ifs * sizeof(uint32_t) );
+	orig_node->bidirect_link = debugMalloc( found_ifs * sizeof(uint16_t), 402 );
+	memset( orig_node->bidirect_link, 0, found_ifs * sizeof(uint16_t) );
 
 	hash_add( orig_hash, orig_node );
